organized_bot/requirements.txt
aiogram>=3.0.0
google-cloud-firestore>=2.11.0
python-dotenv>=1.0.0
pytz>=2023.3
APScheduler>=3.10.1
google-generativeai>=0.3.0 